refactor: remove unused + decompose data collection
