+++ src/backend/SportDataService.Infrastructure/Services/DataCollection/Abstractions/IApiDataService.cs
@@ -1,0 +1,9 @@
+﻿using Newtonsoft.Json.Linq;
+
+namespace SportDataService.Infrastructure.Services.DataCollection.Abstractions;
+
+public interface IApiDataService
+{
+    Task<JToken> GetMarketsDataAsync(CancellationToken cancellationToken);
+    Task<JToken> GetResultsDataAsync(CancellationToken cancellationToken);
+}
