+++ src/backend/SportDataService.Infrastructure/Services/DataCollection/Abstractions/IResultsDataParser.cs
@@ -1,0 +1,13 @@
+﻿using Newtonsoft.Json.Linq;
+using SportDataService.Domain.Models.Results;
+
+namespace SportDataService.Infrastructure.Services.DataCollection.Abstractions;
+
+public interface IResultsDataParser
+{
+    List<TournamentResult> ParseTournamentResultsData(JToken resultsRawData);
+    List<MatchResult> ParseMatchResultsData(JToken resultsRawData);
+    void LinkMatchResultsToTournamentResults(
+        List<TournamentResult> tournamentResults,
+        List<MatchResult> matchResults);
+}
