+++ src/backend/SportDataService.Infrastructure/Services/DataCollection/Helpers/ResultNodeNames.cs
@@ -1,4 +1,4 @@
-﻿namespace Parser.DataCollection;
+﻿namespace SportDataService.Infrastructure.Services.DataCollection.Helpers;
 
 public static class ResultNodeNames
 {
@@ -8,10 +8,10 @@ public static class ResultNodeNames
     public const string MatchOrEventScores = "eventMiscs";
 
     // Competition nodes
-    public const string TournamentId = "id"; 
-    public const string TournamentName = "name"; 
-    public const string SportId = "sportId"; 
-    
+    public const string TournamentId = "id";
+    public const string TournamentName = "name";
+    public const string SportId = "sportId";
+
     // Match record nodes
     public const string MatchOrEventId = "id";
     public const string MatchOrEventName = "name";
@@ -19,12 +19,12 @@ public static class ResultNodeNames
     public const string MatchTournamentId = "competitionId";
     public const string MatchStartTime = "startTime";
     public const string MatchTeam1Id = "team1Id";
-    public const string MatchTeam2Id = "team2Id";    
+    public const string MatchTeam2Id = "team2Id";
     public const string MatchTeam1Name = "team1";
     public const string MatchTeam2Name = "team2";
     public const string MatchOrEventStatus = "status";
     public const string EventMatchId = "parentId";
-    
+
     // Result nodes
     public const string ResultId = "id";
     public const string Team1TotalScore = "score1";
@@ -37,7 +37,7 @@ public static class ResultNodeNames
     public const string MatchTimeCounterType = "timerDirection";
     public const string SecondsPassed = "timerSeconds";
     public const string MatchStartedAt = "timerUpdateTimestampMsec";
-    
+
     // Values
     public const string MatchKindIdValue = "1";
     public const string EventKindIdValue = "104000";
