+++ src/backend/backend.sln
@@ -6,8 +6,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "BettingService", "BettingSe
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SportDataService", "SportDataService", "{F16F9D01-68A9-4EFE-8B4D-2432C7B350C0}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{C4503E4C-7003-455E-8A7E-FB1FBA7A2D60}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Items", "Items", "{197C60B2-11FD-4A83-BF9C-6AE0F79109AC}"
 	ProjectSection(SolutionItems) = preProject
 		stylecop.ruleset = stylecop.ruleset
@@ -30,10 +28,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UserService.Infrastructure", "UserService.Infrastructure\UserService.Infrastructure.csproj", "{DDCB997B-73D1-498A-A16B-F383807472F6}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UserService.Presentation", "UserService.Presentation\UserService.Presentation.csproj", "{6F6435B8-4CA9-4B72-86B0-8D9804508EE2}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTests", "UnitTests\UnitTests.csproj", "{25260583-41F6-4F81-8AC5-D19109F70C14}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IntegrationTests", "IntegrationTests\IntegrationTests.csproj", "{520F4A81-27F9-48C1-9454-2D2AD3C7EE70}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BettingService.DAL", "BettingService.DAL\BettingService.DAL.csproj", "{852EB412-FE7D-48B0-B9D2-2DF9AC67B9E1}"
 EndProject
@@ -59,8 +53,6 @@ Global
 		{329E89F6-7940-48C8-A9A1-36C9EDE2E856} = {6C016A47-8128-457A-9CAC-E4827F68E39B}
 		{DDCB997B-73D1-498A-A16B-F383807472F6} = {6C016A47-8128-457A-9CAC-E4827F68E39B}
 		{6F6435B8-4CA9-4B72-86B0-8D9804508EE2} = {6C016A47-8128-457A-9CAC-E4827F68E39B}
-		{25260583-41F6-4F81-8AC5-D19109F70C14} = {C4503E4C-7003-455E-8A7E-FB1FBA7A2D60}
-		{520F4A81-27F9-48C1-9454-2D2AD3C7EE70} = {C4503E4C-7003-455E-8A7E-FB1FBA7A2D60}
 		{852EB412-FE7D-48B0-B9D2-2DF9AC67B9E1} = {F2174BD1-C0DE-4E16-A6DA-A62852EE8AA2}
 		{B5FF7DEB-3FAB-44EE-8898-F15501E96666} = {F2174BD1-C0DE-4E16-A6DA-A62852EE8AA2}
 		{4865B53A-817C-406B-AE04-A46330CC422A} = {F2174BD1-C0DE-4E16-A6DA-A62852EE8AA2}
@@ -100,14 +92,6 @@ Global
 		{6F6435B8-4CA9-4B72-86B0-8D9804508EE2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6F6435B8-4CA9-4B72-86B0-8D9804508EE2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6F6435B8-4CA9-4B72-86B0-8D9804508EE2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{25260583-41F6-4F81-8AC5-D19109F70C14}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{25260583-41F6-4F81-8AC5-D19109F70C14}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{25260583-41F6-4F81-8AC5-D19109F70C14}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{25260583-41F6-4F81-8AC5-D19109F70C14}.Release|Any CPU.Build.0 = Release|Any CPU
-		{520F4A81-27F9-48C1-9454-2D2AD3C7EE70}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{520F4A81-27F9-48C1-9454-2D2AD3C7EE70}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{520F4A81-27F9-48C1-9454-2D2AD3C7EE70}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{520F4A81-27F9-48C1-9454-2D2AD3C7EE70}.Release|Any CPU.Build.0 = Release|Any CPU
 		{852EB412-FE7D-48B0-B9D2-2DF9AC67B9E1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{852EB412-FE7D-48B0-B9D2-2DF9AC67B9E1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{852EB412-FE7D-48B0-B9D2-2DF9AC67B9E1}.Release|Any CPU.ActiveCfg = Release|Any CPU
